Revert "Double to integer"
This reverts commit 5b940ebbc8af4f55705f07faad7834eddce2bdce.

+++ cult-of-tim-auth/src/main/java/com/cult_of_tim/auth/cultoftimauth/model/User.java
@@ -25,13 +25,13 @@ public class User {
     //@Column(name = "user_email")
     private String email;
 
-    private Integer balance;
+    private double balance;
 
-    public Integer getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(Integer balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
